requirements.txt
Streamlit
PyTorch
Torchvision
PIL
Requests